feat: implement device ban functionality and notifications
- Added `onDeviceBanned` and `notifyDeviceBanned` functions to handle device ban events.
- Updated `registerDeviceLogin` to notify when a device is banned.
- Enhanced i18n support with Arabic and English translations for device ban messages.
- Modified license cache logic to prevent access for banned devices.
- Introduced tests for license cache and expiration reminders, ensuring correct behavior for permanent licenses.
- Implemented UI changes in the Index page to allow editing of contact names with appropriate feedback.
- Added database migrations to enforce device ban checks in validation policies.

+++ supabase/migrations/20260810000001_expiration_authority_and_reminders.sql
@@ -91,7 +91,7 @@ BEGIN
      OR (v_expiry IS NOT NULL AND v_expiry <= v_now) THEN
     v_offline_grace_ms := 0;
   ELSIF v_profile.license_status = 'permanent' THEN
-    v_offline_grace_ms := 3650 * 86400000; -- valid per permanent status
+    v_offline_grace_ms := 3650::bigint * 86400000; -- valid per permanent status
   ELSIF v_expiry IS NULL THEN
     v_offline_grace_ms := 7 * 86400000;    -- undated (malformed/legacy) fallback
   ELSE
